requirements.txt
flask==3.0.3
kubernetes==29.0.0
psycopg2-binary==2.9.9
requests==2.31.0
prometheus-client==0.20.0
